Marked 'process_sequence' as deprecated and moved to pipeline delivery as 'deliver_sequence'
