+++ openpype/lib/delivery.py
@@ -213,6 +213,7 @@ def process_single_file(
     )
 
 
+@deprecated("openpype.pipeline.delivery.deliver_sequence")
 def process_sequence(
     src_path, repre, anatomy, template_name, anatomy_data, format_dict,
     report_items, log
@@ -220,128 +221,33 @@ def process_sequence(
     """ For Pype2(mainly - works in 3 too) where representation might not
         contain files.
 
-        Uses listing physical files (not 'files' on repre as a)might not be
-         present, b)might not be reliable for representation and copying them.
+    Uses listing physical files (not 'files' on repre as a)might not be
+     present, b)might not be reliable for representation and copying them.
 
-         TODO Should be refactored when files are sufficient to drive all
-         representations.
+     TODO Should be refactored when files are sufficient to drive all
+     representations.
 
-        Args:
-            src_path(str): path of source representation file
-            repre (dict): full representation
-            anatomy (Anatomy)
-            template_name (string): user selected delivery template name
-            anatomy_data (dict): data from repre to fill anatomy with
-            format_dict (dict): root dictionary with names and values
-            report_items (collections.defaultdict): to return error messages
-            log (Logger): for log printing
-        Returns:
-            (collections.defaultdict , int)
+    Args:
+        src_path(str): path of source representation file
+        repre (dict): full representation
+        anatomy (Anatomy)
+        template_name (string): user selected delivery template name
+        anatomy_data (dict): data from repre to fill anatomy with
+        format_dict (dict): root dictionary with names and values
+        report_items (collections.defaultdict): to return error messages
+        log (Logger): for log printing
+
+    Returns:
+        (collections.defaultdict , int)
+
+    Deprecated:
+        Function was moved to different location and will be removed
+            after 3.16.* release.
     """
-    src_path = os.path.normpath(src_path.replace("\\", "/"))
 
-    def hash_path_exist(myPath):
-        res = myPath.replace('#', '*')
-        glob_search_results = glob.glob(res)
-        if len(glob_search_results) > 0:
-            return True
-        return False
+    from openpype.pipeline.delivery import deliver_sequence
 
-    if not hash_path_exist(src_path):
-        msg = "{} doesn't exist for {}".format(src_path,
-                                               repre["_id"])
-        report_items["Source file was not found"].append(msg)
-        return report_items, 0
-
-    delivery_templates = anatomy.templates.get("delivery") or {}
-    delivery_template = delivery_templates.get(template_name)
-    if delivery_template is None:
-        msg = (
-            "Delivery template \"{}\" in anatomy of project \"{}\""
-            " was not found"
-        ).format(template_name, anatomy.project_name)
-        report_items[""].append(msg)
-        return report_items, 0
-
-    # Check if 'frame' key is available in template which is required
-    #   for sequence delivery
-    if "{frame" not in delivery_template:
-        msg = (
-            "Delivery template \"{}\" in anatomy of project \"{}\""
-            "does not contain '{{frame}}' key to fill. Delivery of sequence"
-            " can't be processed."
-        ).format(template_name, anatomy.project_name)
-        report_items[""].append(msg)
-        return report_items, 0
-
-    dir_path, file_name = os.path.split(str(src_path))
-
-    context = repre["context"]
-    ext = context.get("ext", context.get("representation"))
-
-    if not ext:
-        msg = "Source extension not found, cannot find collection"
-        report_items[msg].append(src_path)
-        log.warning("{} <{}>".format(msg, context))
-        return report_items, 0
-
-    ext = "." + ext
-    # context.representation could be .psd
-    ext = ext.replace("..", ".")
-
-    src_collections, remainder = clique.assemble(os.listdir(dir_path))
-    src_collection = None
-    for col in src_collections:
-        if col.tail != ext:
-            continue
-
-        src_collection = col
-        break
-
-    if src_collection is None:
-        msg = "Source collection of files was not found"
-        report_items[msg].append(src_path)
-        log.warning("{} <{}>".format(msg, src_path))
-        return report_items, 0
-
-    frame_indicator = "@####@"
-
-    anatomy_data["frame"] = frame_indicator
-    anatomy_filled = anatomy.format(anatomy_data)
-
-    if format_dict:
-        template_result = anatomy_filled["delivery"][template_name]
-        delivery_path = template_result.rootless.format(**format_dict)
-    else:
-        delivery_path = anatomy_filled["delivery"][template_name]
-
-    delivery_path = os.path.normpath(delivery_path.replace("\\", "/"))
-    delivery_folder = os.path.dirname(delivery_path)
-    dst_head, dst_tail = delivery_path.split(frame_indicator)
-    dst_padding = src_collection.padding
-    dst_collection = clique.Collection(
-        head=dst_head,
-        tail=dst_tail,
-        padding=dst_padding
+    return deliver_sequence(
+        src_path, repre, anatomy, template_name, anatomy_data, format_dict,
+        report_items, log
     )
-
-    if not os.path.exists(delivery_folder):
-        os.makedirs(delivery_folder)
-
-    src_head = src_collection.head
-    src_tail = src_collection.tail
-    uploaded = 0
-    for index in src_collection.indexes:
-        src_padding = src_collection.format("{padding}") % index
-        src_file_name = "{}{}{}".format(src_head, src_padding, src_tail)
-        src = os.path.normpath(
-            os.path.join(dir_path, src_file_name)
-        )
-
-        dst_padding = dst_collection.format("{padding}") % index
-        dst = "{}{}{}".format(dst_head, dst_padding, dst_tail)
-        log.debug("Copying single: {} -> {}".format(src, dst))
-        copy_file(src, dst)
-        uploaded += 1
-
-    return report_items, uploaded
